requirements.txt
SpeechRecognition==3.10.0
pywin32==306
openai==0.28.1
PyAutoGUI==0.9.54
psutil==5.9.5
speedtest-cli==2.1.3
twilio==8.5.0 